assignments/assignment4/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
blas=1.0=mkl
bzip2=1.0.6=h1de35cc_5
ca-certificates=2019.1.23=0
cairo=1.14.12=hc4e6be7_4
certifi=2019.3.9=py37_0
cffi=1.12.2=py37hb5b8e2f_1
cycler=0.10.0=py37_0
ffmpeg=4.0=h01ea3c9_0
fontconfig=2.13.0=h5d5b041_1
freetype=2.9.1=hb4e5f40_0
gettext=0.19.8.1=h15daf44_3
glib=2.56.2=hd9629dc_0
graphite2=1.3.13=h2098e52_0
harfbuzz=1.8.8=hb8d4a28_0
hdf5=1.10.2=hfa1e0ec_1
icu=58.2=h4b95b61_1
intel-openmp=2019.3=199
jasper=2.0.14=h636a363_1
jpeg=9b=he5867d9_2
kiwisolver=1.0.1=py37h0a44026_0
libcxx=4.0.1=hcfea43d_1
libcxxabi=4.0.1=hcfea43d_1
libedit=3.1.20181209=hb402a30_0
libffi=3.2.1=h475c297_4
libgfortran=3.0.1=h93005f0_2
libiconv=1.15=hdd342a3_7
libopencv=3.4.2=h7c891bd_1
libopus=1.3=h1de35cc_0
libpng=1.6.36=ha441bb4_0
libtiff=4.0.10=hcb84e12_2
libvpx=1.7.0=h378b8a2_0
libxml2=2.9.9=hab757c2_0
matplotlib=3.0.3=py37h54f8f79_0
mkl=2019.3=199
mkl_fft=1.0.10=py37h5e564d8_0
mkl_random=1.0.2=py37h27c97d8_0
ncurses=6.1=h0a44026_1
ninja=1.9.0=py37h04f5b5a_0
numpy=1.16.2=py37hacdab7b_0
numpy-base=1.16.2=py37h6575580_0
olefile=0.46=py37_0
opencv=3.4.2=py37h6fd60c2_1
openssl=1.1.1b=h1de35cc_1
pcre=8.43=h0a44026_0
pillow=5.4.1=py37hb68e598_0
pip=19.0.3=py37_0
pixman=0.38.0=h1de35cc_0
py-opencv=3.4.2=py37h7c891bd_1
pycparser=2.19=py37_0
pyparsing=2.3.1=py37_0
python=3.7.2=haf84260_0
python-dateutil=2.8.0=py37_0
pytorch=1.0.1=py3.7_2
pytz=2018.9=py37_0
readline=7.0=h1de35cc_5
scipy=1.2.1=py37h1410ff5_0
setuptools=40.8.0=py37_0
six=1.12.0=py37_0
sqlite=3.27.2=ha441bb4_0
tk=8.6.8=ha441bb4_0
torchvision=0.2.2=py_3
tornado=6.0.2=py37h1de35cc_0
wheel=0.33.1=py37_0
xz=5.2.4=h1de35cc_4
zlib=1.2.11=h1de35cc_3
zstd=1.3.7=h5bba6e5_0
